requirements.txt
flet>=0.21.0
sqlcipher3-wheels>=0.5.0
cryptography>=42.0.0